Change IntMetricsObject value type from Int64 to Int
Updated the value property in IntMetricsObject from Int64 to Int and adjusted related code in MatrixValue extension to match the new type. This simplifies type handling and ensures consistency across the codebase.

+++ Sources/Scout/Core/Matrix/MatrixValue.swift
@@ -21,7 +21,7 @@ extension Int: MatrixValue {
     func toObject(in context: NSManagedObjectContext) -> IntMetricsObject {
         let entity = NSEntityDescription.entity(forEntityName: "IntMetricsObject", in: context)!
         let object = IntMetricsObject(entity: entity, insertInto: context)
-        object.value = Int64(self)
+        object.value = self
         return object
     }
 }
